Extract MouseSelectionHandler to own file.
